llm_training/requirements.txt
openai
pandas
python-dotenv
qdrant_client
scikit_learn
sentence_transformers
tqdm
